+++ src/app/components/settings/settings.component.spec.ts
@@ -9,7 +9,8 @@ import { FormModule, SelectComponent, SelectModule, ToolbarModule } from "@batch
 import { ButtonComponent, ButtonsModule } from "@batch-flask/ui/buttons";
 import { PermissionService } from "@batch-flask/ui/permission";
 import { SettingsComponent } from "app/components/settings";
-import { DEFAULT_BE_USER_CONFIGURATION } from "common";
+import { BEUserDesktopConfiguration, DEFAULT_BE_USER_CONFIGURATION } from "common";
+import { updateInput } from "test/utils/helpers";
 
 @Component({
     template: `<bl-settings></bl-settings>`,
@@ -24,7 +25,7 @@ describe("SettingsComponent", () => {
     let resetButtonEl: DebugElement;
     let resetButton: ButtonComponent;
 
-    let settingsServiceSpy: MockUserConfigurationService;
+    let settingsServiceSpy: MockUserConfigurationService<BEUserDesktopConfiguration>;
 
     let themeSelect: SelectComponent;
 
@@ -54,6 +55,7 @@ describe("SettingsComponent", () => {
         resetButton = resetButtonEl.componentInstance;
 
         themeSelect = de.query(By.css("bl-select[formControlName=theme]")).componentInstance;
+
     });
 
     it("disable the reset button when settings are the defaults", () => {
@@ -66,7 +68,7 @@ describe("SettingsComponent", () => {
         expect(resetButton.disabled).toBe(false);
     });
 
-    it("updates the settings", fakeAsync(() => {
+    it("updates the theme", fakeAsync(() => {
         tick(400);
         expect(settingsServiceSpy.current.theme).toEqual("classic");
         themeSelect.selectOption(themeSelect.options.toArray()[1]);
@@ -76,5 +78,31 @@ describe("SettingsComponent", () => {
         expect(settingsServiceSpy.current.theme).toEqual("dark");
         tick(10000);
         expect(settingsServiceSpy.current.theme).toEqual("dark");
+    }));
+
+    it("updates the Microsoft portfolio settings", fakeAsync(() => {
+
+        const repoInput = de.query(By.css("input[formControlName=microsoftPortfolioRepo]"));
+        const branchInput = de.query(By.css("input[formControlName=microsoftPortfolioBranch]"));
+        const pathInput = de.query(By.css("input[formControlName=microsoftPortfolioPath]"));
+
+        expect(repoInput).not.toBeFalsy();
+        expect(branchInput).not.toBeFalsy();
+        expect(pathInput).not.toBeFalsy();
+
+        tick(400);
+        expect(settingsServiceSpy.current.microsoftPortfolio.repo).toEqual("Azure/batch-extension-templates");
+        expect(settingsServiceSpy.current.microsoftPortfolio.branch).toEqual("master");
+        expect(settingsServiceSpy.current.microsoftPortfolio.path).toEqual("templates");
+
+        updateInput(repoInput, "example/my-fork");
+        updateInput(branchInput, "feature/wip");
+        updateInput(pathInput, "foobar");
+
+        tick(400);
+        expect(settingsServiceSpy.current.microsoftPortfolio.repo).toEqual("example/my-fork");
+        expect(settingsServiceSpy.current.microsoftPortfolio.branch).toEqual("feature/wip");
+        expect(settingsServiceSpy.current.microsoftPortfolio.path).toEqual("foobar");
+        tick(10000);
     }));
 });